LayoutTests/http/tests/iframe-monitor/just-use-eligible-subresource-expected.txt

Test iframe won't be unloaded by just using eligible subresource.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is true
PASS document.querySelector('iframe[name=frame1]').srcdoc is ""
PASS successfullyParsed is true

TEST COMPLETE
